Changed Dynamic id & used Xpath
Since id is not unique in this case and is dynamically changing on each page load. Hence we will have to use some other attribute to locate the webelement. I have used Xpath which identifies the webelement uniquey. Also, increased time for waitFor() method So that it could locate element easily.